+++ src/assets/css/modules/_btns.scss
@@ -4,19 +4,23 @@
             margin: 0 1rem 0 0;
             font-size: 1rem;
             border-radius: 2rem;
-            min-width: 8rem;
+            min-width: 10rem;
             text-decoration: none;
             text-align: center;
             cursor: pointer;
-            display: inline-block;
+             @include media-query($on-small) {
+            width: 100%;
+                 margin: 0px;
+                 margin-bottom: 1rem;
+            }
             &:first-child {
                 color: $white;
                 background: $primary;
                 border: 0;
             }
             &:last-child {
-                color: $g2;
+                color: $primary;
                 background: $white;
-                border: 2px solid $g3;
+                border: 2px solid $primary;
             }
         }
