+++ src/assets/css/modules/_home.scss
@@ -1,38 +1,40 @@
-.home {
-    .hero-banner {
-        max-width: 41rem;
-        margin: 3rem auto;
-        @include media-query($on-laptop) {
-            padding: 0 2rem;
-            margin: 0;
-        }
-
-        .row {
-            div:first-child {
-                flex: none;
+    .home {
+        .hero-banner {
+            .row {
+                padding: 140px 0;
+                max-width: 800px;
+                margin: 0 auto;
+                @include media-query($on-laptop) {
+                    max-width: 600px;
+                    .inferno-logo {
+                        display: block;
+                        margin: 0 auto;
+                        max-width: 240px;
+                    }
+                }
+                h1 {
+                    font-weight: bold;
+                    text-align: left;
+                    margin: 0;
+                }
+                h2 {
+                    font-size: 1.6rem;
+                    font-weight: normal;
+                    margin: 0 0 1rem;
+                }
             }
         }
-        h1 {
-            font-weight: bold;
-            text-align: left;
-            margin: 0;
+            section {
+                background: $bg;
+                .row > * {padding: 2rem;
+                outline: 1px solid $white;}
+                h2 {
+                    color: $primary;
+                    border-left: 3px solid $primary;
+                    padding-left: 0.6rem;
+                }
+                p {
+                    padding-left: 0.8rem;
+                }
+            }
         }
-        h2 {
-            font-size: 1.6rem;
-            font-weight: normal;
-            margin: 0 0 1rem;
-        }
-    }
-    section {
-        background: $bg;
-        padding: 2rem;
-        h2 {
-            color: $primary;
-            border-left: 3px solid $primary;
-            padding-left: 0.6rem;
-        }
-        p {
-            padding-left: 0.8rem;
-        }
-    }
-}
