+++ src/components/common/Logo.js
@@ -1,27 +1,19 @@
 import Inferno from 'inferno'
 
 export default ({ width, height, showName = false }) => {
-    const viewBoxHeight = showName ? 305 : 205
     return (
     <svg version="1.0"
          xmlns="http://www.w3.org/2000/svg"
          className="inferno-logo"
          width={width}
          height={height}
-         viewBox={`0 0 300 ${viewBoxHeight}`}
+         viewBox={`0 0 320 320`}
          preserveAspectRatio="xMidYMid meet">
-        <g transform="translate(0.000000,305.000000) scale(0.050000,-0.050000)"
+        <g
            fill="#494949" stroke="none">
-            <path fill="#FF3232" d="M3008 5996 c-524 -341 -715 -834 -528 -1366 135 -383 2 -667 -278
-            -592 -232 62 -326 324 -223 622 47 135 42 142 -34 46 -637 -809 14 -2046 1043
-            -1981 989 63 1514 1231 896 1994 -132 164 -155 170 -112 30 67 -214 11 -484
-            -125 -613 -99 -92 -113 -83 -109 74 6 242 -66 443 -294 823 -260 436 -300 646
-            -174 922 26 58 45 105 42 105 -3 0 -49 -29 -104 -64z"/>
-            <path d="M2390 5877 c-1484 -422 -1946 -2303 -827 -3368 1255 -1196 3318 -309
-            3326 1431 4 934 -564 1694 -1449 1939 -195 54 -197 53 -233 -34 -87 -207 9
-            -517 175 -570 777 -246 1158 -1207 775 -1955 -560 -1095 -2156 -976 -2557 191
-            -196 569 16 1227 509 1579 116 83 126 97 159 214 48 174 162 381 278 509 l98
-            107 -57 -1 c-31 -1 -120 -20 -197 -42z"/>
+            <path fill="#FF3232" d="m 144.94311,258.52282 c -43.42991,-6.10924 -78.349545,-46.3246 -79.249278,-90.01278 -0.713742,-25.25791 8.34156,-51.65879 26.975037,-69.117297 -1.960341,9.972007 -8.403614,22.362757 -6.547263,34.116337 0.03189,14.37259 13.724222,28.46068 28.515044,24.55802 13.25199,-3.19992 18.25555,-18.79679 14.69862,-30.82848 -3.59996,-16.2619 -12.08863,-31.824491 -9.6427,-48.996236 2.36408,-32.356368 26.89643,-58.022237 53.89035,-73.2676739 5.45365,-4.00072606 3.5945,-0.7332729 1.05839,3.2070159 -9.19456,14.874592 -10.49072,34.010739 -2.56271,49.707822 11.50286,27.012559 33.99187,50.392632 35.28169,81.000292 0.2767,6.28473 -1.87795,22.40825 7.34955,11.98701 14.55425,-13.60356 13.45286,-36.23545 6.86038,-53.472975 -1.78943,-9.7282 9.90839,6.576315 12.79534,9.418665 25.34364,31.70617 26.69333,79.92192 2.68761,112.76308 -20.08741,29.17246 -57.19779,44.88699 -92.11006,38.9372 z"/>
+            <path d="M 139.30154,315.44353 C 82.363113,307.67683 34.800143,268.05107 16.436043,213.08158 10.714754,195.95597 9.0805235,185.45803 9.0241835,165.4693 c -0.0352,-12.50176 0.44465,-19.64603 1.8136395,-27 11.04078,-59.309408 54.57033,-106.002753 112.446807,-120.619643 6.27275,-1.584203 12.032,-2.88037 12.79832,-2.88037 2.47413,0 1.49878,2.104191 -3.5215,7.597175 -10.71212,11.720769 -19.3445,27.745026 -22.76755,42.263309 l -1.73541,7.360484 -7.04111,4.780717 C 77.348473,93.041494 61.321623,117.28406 56.003853,145.05941 c -1.88244,9.83218 -2.13576,29.04136 -0.50182,38.05201 7.949831,43.84066 41.671205,78.08395 84.985117,86.30046 16.7819,3.18347 38.60099,1.52687 54.33871,-4.12563 31.78525,-11.41628 56.01144,-36.34418 65.95088,-67.86113 3.71227,-11.77127 4.81236,-19.17255 4.79634,-32.26952 -0.03,-24.54617 -6.95521,-45.18905 -21.75828,-64.85773 -12.86464,-17.093099 -28.5736,-27.906277 -54.48766,-37.506276 -3.27639,-1.213757 -7.30792,-9.70224 -9.42083,-19.83582 -1.62317,-7.784779 -0.9459,-15.17185 2.08454,-22.736486 1.49743,-3.73791 2.08109,-4.249696 4.844,-4.247478 4.85895,0.0039 20.82223,4.585093 31.49825,9.03947 45.63792,19.041595 79.31236,59.862697 89.88242,108.95802 3.25529,15.11998 3.71279,43.13125 0.9473,58 -14.91697,80.20173 -89.52188,134.43295 -169.86128,123.47423 z"
+     style="fill:#494949"/>
             {showName && <InfernoText/>}
         </g>
     </svg>
